+++ Tutorial4Part2.cpp
@@ -14,7 +14,7 @@ int main()
 {
     const int MAX_ROW = 4;
 
-    float mat[MAX_ROW][MAX_COL] = { 1, 2, 3, 4, 5 };
+    double mat[MAX_ROW][MAX_COL] = { 1, 2, 3, 4, 5 };
     int choice, column, row;
 
     do
